tests/wpt/web-platform-tests/app-history/transition/fragment-nav-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    assert_equals(appHistory.transition, null);
    const from = appHistory.current;

    appHistory.addEventListener("navigate", t.step_func(e => {
      e.transitionWhile(Promise.resolve());
      assert_equals(appHistory.transition, null);
    }));

    appHistory.addEventListener("navigatesuccess", t.step_func_done(() => {
      assert_equals(appHistory.transition.navigationType, "push");
      assert_equals(appHistory.transition.from, from);
    }));

    location.href = "#1";

    assert_equals(appHistory.transition.navigationType, "push");
    assert_equals(appHistory.transition.from, from);
  }, 0);
}, "intercepted fragment navigations set transition");
</script>
